--- a/src/Nethermind/Nethermind.Serialization.Rlp/RlpStream.cs
+++ b/src/Nethermind/Nethermind.Serialization.Rlp/RlpStream.cs
@@ -146,7 +146,7 @@ namespace Nethermind.Serialization.Rlp
 
         public virtual void WriteByte(byte byteToWrite)
         {
-            Data[Position++] = byteToWrite;
+            Data[_position++] = byteToWrite;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -157,15 +157,15 @@ namespace Nethermind.Serialization.Rlp
 
         public virtual void Write(ReadOnlySpan<byte> bytesToWrite)
         {
-            bytesToWrite.CopyTo(Data.AsSpan(Position, bytesToWrite.Length));
-            Position += bytesToWrite.Length;
+            bytesToWrite.CopyTo(Data.AsSpan(_position, bytesToWrite.Length));
+            _position += bytesToWrite.Length;
         }
 
         public virtual void Write(IReadOnlyList<byte> bytesToWrite)
         {
             for (int i = 0; i < bytesToWrite.Count; ++i)
             {
-                Data![Position + i] = bytesToWrite[i];
+                Data![_position + i] = bytesToWrite[i];
             }
             Position += bytesToWrite.Count;
         }
@@ -175,7 +175,19 @@ namespace Nethermind.Serialization.Rlp
 
         public byte[]? Data { get; }
 
-        public virtual int Position { get; set; }
+        private int _position = 0;
+
+        public virtual int Position
+        {
+            get
+            {
+                return _position;
+            }
+            set
+            {
+                _position = value;
+            }
+        }
 
         public virtual int Length => Data!.Length;
 
@@ -672,6 +684,13 @@ namespace Nethermind.Serialization.Rlp
 
         public (int PrefixLength, int ContentLength) ReadPrefixAndContentLength()
         {
+            (int prefixLength, int contentLength) result = PeekPrefixAndContentLength();
+            Position += result.prefixLength;
+            return result;
+        }
+
+        public (int PrefixLength, int ContentLength) PeekPrefixAndContentLength()
+        {
             (int prefixLength, int contentLength) result;
             int prefix = ReadByte();
             if (prefix <= 128)
@@ -691,7 +710,7 @@ namespace Nethermind.Serialization.Rlp
                     throw new RlpException("Expected length of length less or equal 4");
                 }
 
-                int length = DeserializeLength(lengthOfLength);
+                int length = PeekDeserializeLength(1, lengthOfLength);
                 if (length < 56)
                 {
                     throw new RlpException($"Expected length greater or equal 56 and was {length}");
@@ -706,7 +725,7 @@ namespace Nethermind.Serialization.Rlp
             else
             {
                 int lengthOfContentLength = prefix - 247;
-                int contentLength = DeserializeLength(lengthOfContentLength);
+                int contentLength = PeekDeserializeLength(1, lengthOfContentLength);
                 if (contentLength < 56)
                 {
                     throw new RlpException($"Expected length greater or equal 56 and got {contentLength}");
@@ -716,15 +735,6 @@ namespace Nethermind.Serialization.Rlp
                 result = (lengthOfContentLength + 1, contentLength);
             }
 
-            return result;
-        }
-
-        public (int PrefixLength, int ContentLength) PeekPrefixAndContentLength()
-        {
-            int memorizedPosition = Position;
-            (int PrefixLength, int ContentLength) result = ReadPrefixAndContentLength();
-
-            Position = memorizedPosition;
             return result;
         }
 
@@ -764,6 +774,26 @@ namespace Nethermind.Serialization.Rlp
             // additional bounds checking from BinaryPrimitives.ReadUInt16BigEndian etc
             ref byte firstElement = ref MemoryMarshal.GetReference(Read(lengthOfLength));
 
+            return DeserializeLengthRef(ref firstElement, lengthOfLength);
+        }
+
+        private int PeekDeserializeLength(int offset, int lengthOfLength)
+        {
+            if (lengthOfLength == 0 || (uint)lengthOfLength > 4)
+            {
+                ThrowArgumentOutOfRangeException(lengthOfLength);
+            }
+
+            // Will use Unsafe.ReadUnaligned as we know the length of the span is same
+            // as what we asked for and then explicitly check lengths, so can skip the
+            // additional bounds checking from BinaryPrimitives.ReadUInt16BigEndian etc
+            ref byte firstElement = ref MemoryMarshal.GetReference(Peek(offset, lengthOfLength));
+
+            return DeserializeLengthRef(ref firstElement, lengthOfLength);
+        }
+
+        private static int DeserializeLengthRef(ref byte firstElement, int lengthOfLength)
+        {
             int result = firstElement;
             if (result == 0)
             {
@@ -818,27 +848,27 @@ namespace Nethermind.Serialization.Rlp
             {
                 throw new RlpException("Length starts with 0");
             }
+        }
 
-            [DoesNotReturn]
-            static void ThrowArgumentOutOfRangeException(int lengthOfLength)
-            {
-                throw new InvalidOperationException($"Invalid length of length = {lengthOfLength}");
-            }
+        [DoesNotReturn]
+        static void ThrowArgumentOutOfRangeException(int lengthOfLength)
+        {
+            throw new InvalidOperationException($"Invalid length of length = {lengthOfLength}");
         }
 
         public virtual byte ReadByte()
         {
-            return Data![Position++];
+            return Data![_position++];
         }
 
         public virtual byte PeekByte()
         {
-            return Data![Position];
+            return Data![_position];
         }
 
         protected virtual byte PeekByte(int offset)
         {
-            return Data![Position + offset];
+            return Data![_position + offset];
         }
 
         protected virtual void SkipBytes(int length)
@@ -848,7 +878,7 @@ namespace Nethermind.Serialization.Rlp
 
         public virtual Span<byte> Read(int length)
         {
-            Span<byte> data = Data.AsSpan(Position, length);
+            Span<byte> data = Data.AsSpan(_position, length);
             Position += length;
             return data;
         }
@@ -1029,9 +1059,12 @@ namespace Nethermind.Serialization.Rlp
 
         public Span<byte> Peek(int length)
         {
-            Span<byte> item = Read(length);
-            Position -= item.Length;
-            return item;
+            return Peek(0, length);
+        }
+
+        public virtual Span<byte> Peek(int offset, int length)
+        {
+            return Data.AsSpan(_position + offset, length);
         }
 
         public bool IsNextItemEmptyArray()