## Minor rlpstream improvement

- Improve PatriciaTree benchmark by about 9%.
- Works by:
  - `PeekPrefixAndContentLength` without changing `Position`.
  - Use `_position` instead of virtual property `Position` whenever it is safe to do so (array backed RlpStream).

|   Method |     Mean |   Error |  StdDev |   Gen0 |   Gen1 | Allocated |
|--------- |---------:|--------:|--------:|-------:|-------:|----------:|
| Before | 164.2 us | 1.55 us | 2.32 us | 1.7090 | 0.4883 | 152.98 KB |
| After | 149.4 us | 1.55 us | 2.32 us | 1.7090 | 0.4883 | 152.98 KB |

## Changes

- Make `PeekPrefixAndContentLength` not change `Position`.
- Use `_position` instead of virtual property `Position` whenever safe.

## Types of changes

#### What types of changes does your code introduce?

- [X] Optimization
- [ ] Other: _Description_

## Testing

#### Requires testing

- [X] Yes
- [ ] No

#### If yes, did you write tests?

- [ ] Yes
- [X] No

#### Notes on testing

- Existing unit tests passed.
- Can sync and process block as usual.